[#13]Refactor: move login authorization logic in interceptor of axios instance to useAuthApi hook
